\name{mixcat-package}

\alias{mixcat-package}
\alias{mixcat}

\docType{package}

\title{MIXED EFFECTS REGRESSION MODELS FOR MULTICATEGORY DATA}

\description{Mixed effects models for the analysis of binary or
             multinomial (ordinal or nominal) data with non-parametric
             distribution for the random effects.
             The main function  is \code{\link{npmlt}} and it fits
             baseline or cumulative logit models.}

\details{
\tabular{ll}{
Package: \tab mixcat\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2011-09-11\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab no\cr}

This program is free software; you can redistribute it and/or
modify it under the terms of the GNU General Public License
as published by the Free Software Foundation; either version 2
of the License, or (at your option) any later version.

This program is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

For details on  the GNU General Public License
see \url{http://www.gnu.org/copyleft/gpl.html} or
write to the Free Software Foundation, Inc., 51 Franklin Street,
Fifth Floor, Boston, MA  02110-1301, USA.}

\author{
Georgios Papageorgiou and John Hinde (2011)

Maintainer: Georgios Papageorgiou <gpapageo@gmail.com>}

\section{Acknowledgments}{
Papageorgiou's work on this \R package was supported by Science
Foundation Ireland Research Frontiers grant 07/RFP/MATF448.}

\references{
Papageorgiou, G. and Hinde, J. (2010).
Multivariate generalized linear mixed models with semi-nonparametric and smooth
nonparametric random effects densities. Available online: Statistics and Computing}

\keyword{models}
\keyword{regression}

\seealso{ \code{\link[npmlreg]{npmlreg}}}
